Buffer cards instead of loading all at once
While using this project, I found when putting on larger info to the
cards, such as UIImages, performance was greatly affected. I fixed this
by only loading a set number of cards a time.

ALSO: Created a method in ViewController that acts as a delegate once a
card is swiped